updated settings page
