+++ frontend/src/lib/components/navbar/navbar.svelte
@@ -141,7 +141,7 @@
           </a>
           <Navlink href="/dashboard" class="text-muted-foreground hover:text-foreground"> Dashboard </Navlink>
           <Navlink href="/recipes" class="text-muted-foreground hover:text-foreground"> Recipes </Navlink>
-          <Navlink href="/settings" class="hover:text-foreground"> Settings </Navlink>
+          <Navlink href="/settings/general" class="hover:text-foreground"> Settings </Navlink>
         </nav>
       </Sheet.Content>
     </Sheet.Root>
@@ -305,7 +305,7 @@
         <DropdownMenu.Content align="end" class="w-56">
           <DropdownMenu.Label class="font-medium">My Account</DropdownMenu.Label>
           <DropdownMenu.Separator class="my-1" />
-          <DropdownMenu.Item onclick={() => goto('/settings')} class="cursor-pointer"><Cog />Settings</DropdownMenu.Item>
+          <DropdownMenu.Item onclick={() => goto('/settings/general')} class="cursor-pointer"><Cog />Settings</DropdownMenu.Item>
           <DropdownMenu.Separator class="my-1" />
           <DropdownMenu.Item onclick={handleLogout} class="cursor-pointer"><LogOut />Log out</DropdownMenu.Item>
         </DropdownMenu.Content>
